requirements.txt
opencv-python
numpy
tensorflow
matplotlib
keras
pywin32
Pillow
scikit-image
ghostscript